Add files via upload
I initially created this program as an exercise which was a part of a course I am following. However, I got a little carried away and had it so it actually saved the things to do and made it more visually appealing as command prompt isn't the most exciting thing to look at.